+++ src/NyanCat/PHPUnit/ResultPrinter.php
@@ -11,6 +11,13 @@
 
 namespace NyanCat\PHPunit;
 
+use NyanCat\Cat;
+use NyanCat\Rainbow;
+use NyanCat\Team;
+use NyanCat\Scoreboard;
+
+use Fab\Factory as FabFactory;
+
 /**
  * Mmmm poptarts...
  *
@@ -24,200 +31,41 @@ namespace NyanCat\PHPunit;
 class ResultPrinter extends \PHPUnit_TextUI_ResultPrinter
 {
     /**
-     * Ansi escape
-     */
-    const ESC = "\x1b[";
-    /**
-     * Ansi Normal
-     */
-    const NND = "\x1b[0m";
-
-    /**
-     * The Fab to paint the rainbow with.
+     * The Nyan Cat scoreboard.
      *
-     * @var Fab\Fab
+     * @var NyanCat\Scoreboard
      */
-    private $fab;
-
+    private $scoreboard;
     /**
-     * For the shakes
+     * The Nyan Cat rainbow.
      *
-     * @var integer
-     */
-    private $tick = 0;
-
-    /**
-     * Length of the rainbow
-     *
-     * @var integer
-     */
-    private $trajectoryWidthMax;
-    /**
-     * Lines in the rainbow
-     *
-     * @var integer
-     */
-    private $numberOfLines = 4;
-    /**
-     * Length of the cat
-     *
-     * @var integer
-     */
-    private $nyanCatWidth = 11;
-    /**
-     * Length of the scoreboard
-     *
-     * @var integer
-     */
-    private $scoreboardWidth = 5;
-
-    /**
-     * Holds the rainbow chars
-     *
-     * @var array
-     */
-    private $trajectories = array();
-
-    /**
-     * Status colors
-     *
-     * @var array
-     */
-    private static $paints = array(
-        'passes'   => 32,
-        'failures' => 31,
-        'pending'  => 36
-    );
-
-    /**
-     * Statuses
-     *
-     * @var array
-     */
-    private $stats = array(
-        'passes'   => 0,
-        'failures' => 0,
-        'pending'  => 0
-    );
+     * @var NyanCat\Scoreboard
+     */    
+    private $rainbow;
 
     /**
      * {@inheritdoc}
      */
     public function __construct($out = NULL, $verbose = FALSE, $colors = FALSE, $debug = FALSE)
     {
-        $term = empty($_SERVER['TERM']) ? 'unknown' : $_SERVER['TERM'];
-        $this->fab = \Fab\Factory::getFab($term);
-
-        for ($i = 0; $i < $this->numberOfLines; $i++) {
-            $this->trajectories[] = array();
-        }
-
-        parent::__construct($out, $verbose, true, $debug);
-    }
-
-    /**
-     * Moves the cursor up a given number of lines
-     *
-     * @param integer $lines The number of lines to move the cursor
-     */
-    protected function cursorUp($lines)
-    {
-        $this->write(self::ESC . $lines . 'A');
-    }
-
-    /**
-     * Adds a new color segment to the rainbow.
-     */
-    protected function appendRainbow()
-    {
-        $segment = $this->fab->paint(
-            $this->tick ? '_' : '-'
+        $this->rainbow = new Rainbow(
+            FabFactory::getFab(
+                empty($_SERVER['TERM']) ? 'unknown' : $_SERVER['TERM']
+            )
+        );
+        $this->scoreboard = new Scoreboard(
+            new Cat(),
+            $this->rainbow,
+            array(
+                new Team('pass', 'green', '^'),
+                new Team('fail', 'red', 'o'),
+                new Team('pending', 'cyan', '-'),
+            ),
+            5,
+            array($this, 'write')
         );
 
-        foreach ($this->trajectories as &$trajectory) {
-            if (count($trajectory) >= $this->trajectoryWidthMax) {
-                array_shift($trajectory);
-            }
-
-            $trajectory[] = $segment;
-        }
-    }
-
-    /**
-     * Draws the scoreboard for the current iteration.
-     */
-    protected function drawScoreboard()
-    {
-        foreach ($this->stats as $key => $stat) {
-            $this->write(' ');
-            $this->write(
-                self::ESC . self::$paints[$key] . 'm' . $stat . self::NND
-            );
-            $this->write("\n");
-        }
-
-        $this->write("\n");
-        $this->cursorUp($this->numberOfLines);
-    }
-
-    /**
-     * Draws the rainbow for the current iteration.
-     */
-    protected function drawRainbow(){
-
-        foreach ($this->trajectories as $line) {
-            $this->write(self::ESC . $this->scoreboardWidth . 'C');
-            $this->write(implode($line));
-            $this->write("\n");
-        }
-
-        $this->cursorUp($this->numberOfLines);
-    }
-
-    /**
-     * Draws the nyan cat for a given status.
-     *
-     * @param string $status The status (pass/fail/etc)
-     */
-    protected function drawNyanCat($status)
-    {
-        $startWidth = $this->scoreboardWidth + count($this->trajectories[0]);
-
-        // Back
-        $this->write(self::ESC . $startWidth . 'C');
-        $this->write('_,------,');
-        $this->write("\n");
-
-        // Ears
-        $padding = $this->tick ? '  ' : '   ';
-        $this->write(self::ESC . $startWidth . 'C');
-        $this->write('_|' . $padding . '/\\_/\\ ');
-        $this->write("\n");
-
-        // Face
-        $padding = $this->tick ? '_' : '__';
-        $tail = $this->tick ? '~' : '^';
-        switch ($status) {
-            case 'pass':
-                $face = '( ^ .^)';
-                break;
-            case 'fail':
-                $face = '( o .o)';
-                break;
-            default:
-                $face = '( - .-)';
-        }
-        $this->write(self::ESC . $startWidth . 'C');
-        $this->write($tail . '|' . $padding . $face . ' ');
-        $this->write("\n");
-
-        // Feet
-        $padding = $this->tick ? ' ' : '  ';
-        $this->write(self::ESC . $startWidth . 'C');
-        $this->write($padding . '""  "" ');
-        $this->write("\n");
-
-        $this->cursorUp($this->numberOfLines);
+        parent::__construct($out, $verbose, true, $debug);
     }
 
     /**
@@ -229,11 +77,7 @@ class ResultPrinter extends \PHPUnit_TextUI_ResultPrinter
             return parent::writeProgress($progress);
         }
 
-        $this->appendRainbow();
-        $this->drawScoreboard();
-        $this->drawRainbow();
-        $this->drawNyanCat($progress);
-        $this->tick = !$this->tick;
+        $this->scoreboard->score($progress);
     }
 
     /**
@@ -242,10 +86,10 @@ class ResultPrinter extends \PHPUnit_TextUI_ResultPrinter
     protected function printHeader()
     {
         if (!$this->debug) {
-            if (count($this->trajectories[0]) === 0) {
-                $this->writeProgress('end');
+            if ($this->rainbow->getWidth() === 0) {
+                $this->scoreboard->start();
             }
-            $this->write("\n\n\n\n");
+            $this->scoreboard->end();
         }
 
         parent::printHeader();
@@ -260,7 +104,6 @@ class ResultPrinter extends \PHPUnit_TextUI_ResultPrinter
             return parent::addError($test, $e, $time);
         }
 
-        $this->stats['failures']++;
         $this->writeProgress('fail');
         $this->lastTestFailed = TRUE;
     }
@@ -274,7 +117,6 @@ class ResultPrinter extends \PHPUnit_TextUI_ResultPrinter
             return parent::addFailure($test, $e, $time);
         }
 
-        $this->stats['failures']++;
         $this->writeProgress('fail');
         $this->lastTestFailed = TRUE;
     }
@@ -288,7 +130,6 @@ class ResultPrinter extends \PHPUnit_TextUI_ResultPrinter
             return parent::addIncompleteTest($test, $e, $time);
         }
 
-        $this->stats['pending']++;
         $this->writeProgress('pending');
         $this->lastTestFailed = TRUE;
     }
@@ -302,7 +143,6 @@ class ResultPrinter extends \PHPUnit_TextUI_ResultPrinter
             return parent::addSkippedTest($test, $e, $time);
         }
 
-        $this->stats['pending']++;
         $this->writeProgress('pending');
         $this->lastTestFailed = TRUE;
     }
@@ -318,8 +158,7 @@ class ResultPrinter extends \PHPUnit_TextUI_ResultPrinter
 
         if ($this->numTests == -1) {
             parent::startTestSuite($suite);
-            $this->trajectoryWidthMax = $this->maxColumn - $this->nyanCatWidth;
-            $this->writeProgress('start');
+            $this->scoreboard->start();
         }
     }
 
@@ -333,7 +172,6 @@ class ResultPrinter extends \PHPUnit_TextUI_ResultPrinter
         }
 
         if (!$this->lastTestFailed) {
-            $this->stats['passes']++;
             $this->writeProgress('pass');
         }