+++ ChangeLog.md
@@ -7,3 +7,4 @@ Nyan Cat result printer for PHPUnit 1.0.2
 -----------------------------------------
 
 * Fixed #2: Added support for `--debug` and `--verbose`. (This requires PHPUnit 3.8+)
+* Use `whatthejeff/nyancat-scoreboard` for scoreboard generation.